chore: cleaning type exports
