+++ packages/uikit-react-native/src/localization/StringSet.en.ts
@@ -1,6 +1,6 @@
 import dateLocale from 'date-fns/locale/en-US';
 
-import { createBaseStringSet } from './StringSet.type';
+import { createBaseStringSet } from './createBaseStringSet';
 
 const StringSetEn = createBaseStringSet({ dateLocale });
